few changes to Savings Account future value method

+++ src/main/java/com/meritamerica/assignment1/SavingsAccount.java
@@ -48,7 +48,7 @@ public class SavingsAccount extends AccountHolder
 	  
 	double futureValue(int years) 
 	{	 
-		return currentBalance * Math.pow((1 + interestRate), years);
+		return currentBalance * Math.pow((1 + interestRate), (double)years);
 	}
 	
 	public String toString()
